made both routes and have jwt

+++ src/routes/events.route.js
@@ -2,6 +2,8 @@ const express = require("express");
 const router = express.Router();
 const eventCreatorModel = require("../models/events.model");
 const wrapAsync = require("../utils/wrapAsync");
+const { protectRoute } = require("../middleware/auth");
+const uuidv4 = require("uuid/v4");
 
 const requireJsonContent = (req, res, next) => {
   if (req.headers["content-type"] !== "application/json") {
@@ -26,21 +28,39 @@ const getSingleEvent = async (req, res) => {
 const createEvent = async (req, res) => {
   const event = new eventCreatorModel(req.body);
   await eventCreatorModel.init();
+  event.eventOwner = req.user.username;
+  event.eventId = uuidv4();
   const newEvent = await event.save();
   res.status(201).send(newEvent);
 };
 
-const editSingleEvent = async (req, res) => {
+const editSingleEvent = async (req, res, next) => {
   const newEvent = req.body;
-  const foundEvent = await eventCreatorModel.findOneAndUpdate(
+  const eventOwnerCheck = await eventCreatorModel.findOne({
+    eventId: req.params.id
+  });
+  if (eventOwnerCheck.eventOwner != req.user.username) {
+    const err = new Error("You cannot edit as this is not your post.");
+    err.statusCode = 403;
+    next(err);
+  }
+  const updatedEvent = await eventCreatorModel.findOneAndUpdate(
     { eventId: req.params.id },
     newEvent,
     { new: true }
   );
-  res.status(200).send(foundEvent);
+  res.status(200).send(updatedEvent);
 };
 
 const deleteSingleEvent = async (req, res) => {
+  const eventOwnerCheck = await eventCreatorModel.findOne({
+    eventId: req.params.id
+  });
+  if (eventOwnerCheck.eventOwner != req.user.username) {
+    const err = new Error("You cannot delete THIS SHIT.");
+    err.statusCode = 403;
+    next(err);
+  }
   const deletedEvent = await eventCreatorModel.findOneAndDelete({
     eventId: req.params.id
   });
@@ -49,9 +69,14 @@ const deleteSingleEvent = async (req, res) => {
 
 router.get("/", wrapAsync(getAllEvents));
 router.get("/:id", wrapAsync(getSingleEvent));
-router.post("/create", requireJsonContent, wrapAsync(createEvent));
-router.patch("/:id", wrapAsync(editSingleEvent));
-router.delete("/:id", wrapAsync(deleteSingleEvent));
+router.post(
+  "/create",
+  requireJsonContent,
+  protectRoute,
+  wrapAsync(createEvent)
+);
+router.patch("/:id", protectRoute, wrapAsync(editSingleEvent));
+router.delete("/:id", protectRoute, wrapAsync(deleteSingleEvent));
 
 router.use((err, req, use, next) => {
   if (err.name === "ValidationError") {
